+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### v0.1.2 (Jule 5, 2021)
+
+- **Changes**:
+  - Add `thiserror` crate to dependencies.
+  - Remove `anyhow` crate from dependencies.
+  - Refactor `utils::to_map`.
+  - Remove useless trait requirements.
+  - Add simple error handling.
+
 ### v0.1.1 (Jule 4, 2021)
 
 - **Fixes**: